backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-wildfire-tracker"
version = "0.1.0"
description = "AI-powered wildfire detection, tracking, and risk assessment"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "AI Wildfire Tracker Team" },
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "duckdb>=0.10.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.6.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.26.0",
    "ruff>=0.9.0",
    "mypy>=1.13.0",
    "pandas-stubs>=2.2.0",
    "types-requests>=2.31.0",
    "pytest-json-report>=1.5.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

# ── Ruff (linter + formatter) ──────────────────────────────────────────────

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (formatter handles this)
]

[tool.ruff.lint.isort]
known-first-party = ["ai_wildfire_tracker"]

[tool.ruff.format]
quote-style = "double"

# ── Pytest ──────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

# ── Mypy ────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
